+++ src/vector/defn.rs
@@ -69,16 +69,47 @@ impl Defn {
 
     /// Get the index of a field.
     ///
+    /// The comparison is done case-insensitively, and if multiple fields match the requested
+    /// name, the first one is returned.
     /// If the field is missing, returns [`GdalError::InvalidFieldName`].
     ///
     pub fn field_index<S: AsRef<str>>(&self, field_name: S) -> Result<usize> {
-        let c_str_field_name = CString::new(field_name.as_ref())?;
+        self._field_index(field_name.as_ref())
+    }
+
+    fn _field_index(&self, field_name: &str) -> Result<usize> {
+        let c_str_field_name = CString::new(field_name)?;
         let field_idx =
             unsafe { gdal_sys::OGR_FD_GetFieldIndex(self.c_defn(), c_str_field_name.as_ptr()) };
         if field_idx == -1 {
             return Err(GdalError::InvalidFieldName {
-                field_name: field_name.as_ref().to_string(),
+                field_name: field_name.to_string(),
                 method_name: "OGR_FD_GetFieldIndex",
+            });
+        }
+
+        let idx = field_idx.try_into()?;
+        Ok(idx)
+    }
+
+    /// Get the index of a geometry field.
+    ///
+    /// The comparison is done case-insensitively, and if multiple fields match the requested
+    /// name, the first one is returned.
+    /// If the field is missing, returns [`GdalError::InvalidFieldName`].
+    ///
+    pub fn geometry_field_index<S: AsRef<str>>(&self, field_name: S) -> Result<usize> {
+        self._geometry_field_index(field_name.as_ref())
+    }
+
+    fn _geometry_field_index(&self, field_name: &str) -> Result<usize> {
+        let c_str_field_name = CString::new(field_name)?;
+        let field_idx =
+            unsafe { gdal_sys::OGR_FD_GetGeomFieldIndex(self.c_defn(), c_str_field_name.as_ptr()) };
+        if field_idx == -1 {
+            return Err(GdalError::InvalidFieldName {
+                field_name: field_name.to_string(),
+                method_name: "OGR_FD_GetGeomFieldIndex",
             });
         }